feat: add initialize script
