+++ package.json
@@ -96,6 +96,7 @@
     "deploy:dev:fixture": "hardhat deploy --network localhost --tags fixture",
     "deploy:rskdev": "hardhat deploy --network rskdev --tags everything",
     "devchain:start": "ts-node ./scripts/start-dev-chain.ts",
-    "deploy:rskTestnetMocked": "hardhat deploy --network rskTestnetMocked --tags everything"
+    "deploy:rskTestnetMocked": "hardhat deploy --network rskTestnetMocked --tags everything",
+    "initialize:rskTestnetMocked": "hardhat initialize --network rskTestnetMocked"
   }
 }
